Replace pairwise bilinear layer for labels to bilinear
Most biaffine parsers use a pairwise bilinear layer to predict labels,
giving labels tensors for the form:

[batch_size, seq_len, seq_len, n_labels]

The parser then selects the predicted heads to get a tensor of the
shape:

[batch_size, seq_len, n_labels]

And during backpropagation, loss gradients are only backproped into the
vectors corresponding to the active heads. I guess this approach is
taken because it fits in the same framework as pairwise bilinear
predictions for heads. However, it seems like a waste of compute and
memory to compute a quadratic tensor of label weights, only to discard
most of them.

This commit tries to get what we want more directly. We take the token
representations:

tokens: [batch_size, seq_len, embed_size]

And the head representations:

token_heads = tokens[token_heads]

Then we use a (non-pairwise) bilinear form to get label predictions of
the form:

[batch_size, seq_len, n_labels]

This approach still provides the interaction between the dependent and
head representations, but avoids the quadratic factor.

Fixes #4.