Store nI for (Pairwise)Bilinear
In contrast to Thinc models, the shapes of PyTorch models are not stored
in the serialized model. When loading a pipe, we cannot get the shape
output shape of the tok2vec component when it is uses a transformer.

Work around this issues by storing the nI dimenion in the serialized
pipe configuration.

Fixes #20.