Coref Components (#17)
* Add files copied directly from spaCy

These come from the feature/coref branch.

* Add coref tests

These are copied from the feature/coref branch

* Fix relative imports

* Move registry stuff to setup.cfg

* Change default config to use spacy-experimental

* Add init.py files

* Fix imports in tests

* Remove unused class

* Update scorer code

* Fix component names

Tests work now

* Import mean

* Add return type, clean up old TODO comments

* Add type import

* Rename span predictor to span resolver

* Remove error reference

The EXXX assignment will have to be made later.

* Black formatting

* Cleanup from code review

* Fix docstring

* Formatting

* Remove default values from model builder functions

These should be provided by the default config.

* Raise spaCy pin

* Raise torch version used in tests

The former version was old and not compatible with the numpy version
installed, resulting in dlpack related errors.

* Feedback from review

* Update spacy_experimental/coref/span_resolver_component.py

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Apply suggestions from code review

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Change scorers to use direct parameters instead of config

Also assigns "coref_head_clusters" to a var in coref_util for
consistency.

* More scorer fixes

* Minor type related fixes

* More scorer cleanup

* Rename _initialize_from_disk to _initialize_before_deserializing

* Update spacy_experimental/coref/span_resolver_component.py

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Use variables for cluster names in span resolver

* Rename SpanResolver(nn.Module) to SpanResolverModule

* Update error message about existing spans

* Use module-level skips in tests

* Enumerate output clusters from 1 instead of 0

* scorer fix

* Remove old FIXME and document

The code was already correct behaving correctly, just had to track down
why the case was occurring.

* Add warnings for pre-existing spans

I'm not sure this is the right way to do this, but I wanted to try
writing it to show what it would look like for discussion.

* Standardize prefix check in coref

Prefix code didn't handle cases with near matches and was implemented
separately in multiple locations. This uses a single function to match
strictly defined valid span keys.

* Remove old todos

These are sizes that were checked in development but the note was not
removed, or architecture comments that are no longer relevant.

* Update scorers

* Add end-to-end custom label tests

This adds a test for each coref component that uses non-default labels
and tests both that the component/model label document correctly and the
the scorer is able to score correctly.

* Linting and cleanup

* Fix test

Accidentally removed too much when linting

* Fix tests

The config is not a format string, so it should have a literal value.

Also remove some leftover print statements.

* Use function to generate training data with custom prefix

This is just for the cluster tests, if this looks OK I can do the same
in the spanresolver tests.

* Use pytest fixtures for train data

* Use pytest fixtures for train data in span resolver tests

* Use relative imports in tests

* Revert "Use relative imports in tests"

This reverts commit 5b2d6a62f991128377d24362b2b29912c9d8dcbf.

* Remove special handling of blank data in initialization

Since the main initialization was moved for dimension inference, there's
no need to make sure there are annotations on the examples here.

* Remove old TODO

* Remove old TODO

On reflection, the current architecture is fine.

* Remove nI argument from coref model

This isn't necessary because it will always be set when doing dimension
inference. The presence as an argument is a leftover from testing.

* Clean up error/warning strings

* Remover superfluous parametrization

* Clarify why the warning was ignored and narrow scope

* Add note about coref in README

This just points to the main spaCy docs. I don't think we need more
detail here.

* Remove warning about existing annotations

After discussion it was decided to just use the error when an actual
overwrite happens.

* `Coref`: Optimize `create_gold_scores`

Copy the entire mentions array to the CPU, create tuple keys on-demand, pre-allocate output matrix as `Float2d`. Also remove unnecessary casts in `CoreferenceResolver.update`.

* Add Span Cleaner component

This is a component to remove spans based on a prefix. It needs tests
still.

* Remove superfluous serialization code

* Remove unused function

* Apply suggestions from code review

Co-authored-by: kadarakos <kadar.akos@gmail.com>

* Add types, fix formatting

* Clean up get_clusters_from_doc based on review

* Add test for span cleaner

* `SpanResolver`: Optimize `set_annotations`

Coerce scalar tensors to native Python integers to avoid comparison overhead.

* Vectorize `get_predicted_antecedents`

* Add small test of get_predicted_antecedents

I believe this fails on Windows due to type issues.

* Fix dtype issue

The array created by full defaults to int32, while the input is int64.
Linux happily converts this without issue, but Windows throws an error.
Making sure the dtype matches resolves the issue.

* Comment to help understand why warning happens

* Add Windows to the test matrix

No idea if this'll work, but want to check.

* Try using bash

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>
Co-authored-by: shademe <shadeMe@users.noreply.github.com>
Co-authored-by: kadarakos <kadar.akos@gmail.com>