Fix issue with resolving final token in SpanResolver
The SpanResolver seems unable to include the final token in a Doc in
output spans. It will even produce empty spans instead of doing so.

This makes changes so that within the model span end indices are treated
as inclusive, and converts them back to exclusive when annotating docs.
This has been tested to work, though an automated test should be added.