Fix handling of small docs in coref (#28)
* Fix handling of small docs in coref

Docs with one or zero tokens fail in the coref component. This doesn't
have a fix yet, just a failing test. (There is also a test for the span
resolver, which does not fail.)

* Add example short doc to tests

It might be better to include this optionally? On the other hand, since
it should just be ignored in training, having it always there is more
thorough.

* Skip short docs

There can be no coref prediction for docs with one token (or no tokens).
Attempting to treat docs like that normally causes a mess with size
inference, so instead they're skipped.

In training, this just involves skipping the docs in the update step.
This is simple due to the fake batching structure, since the batch
doesn't have to be maintained.

In inference, this just involves short-circuiting to an empty
prediction.

* Clean up retokenization test

The retokenization test is hard-coded to the the training example
because it manually merges some tokens, to make sure that the prediction
and merge line up. It would probably be better to separate out the
training data from the general example here, but for now narrowing the
training data works.