Support lazy, recursive sentence splitting (#40)
* Support lazy, recursive sentence splitting

We use sentence splitting in the biaffine parser to keep the O(n^2)
biaffine attention model tractable. However, since the sentence splitter
makes errors, the parser may not have the correct head available.

This change adds another splitting strategy. The goal of this strategy
is to use the highest-probability splits to partition a doc until each
partition is smaller than or equal to a maximum length. This reduces
the number of attachment errors as a result of incorrect sentence
splits, while providing an upper bound on complexity.

The algorithm works as follows:

* If the length |d| > max_length:
  - Find the highest-probability split in d according to senter.
  - Split d into d_1 and d_2 using the highest probability split.
  - Recursively apply this algorithm to d_1 and d_2.
* Otherwise: do nothing

* `max_tokens` -> `max_length`

* Remove unused `is_train` argument from `split_lazily`

* Do not try to get the senter pipe

* Rename `senter` option to `senter_name`

* Fix example project

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* Formatting fixes

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* ArcPredicter.__init__:  `Language` -> `Vocab` arg

* Update spacy_experimental/biaffine_parser/arc_predicter.pyx

---------

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>